requirements.txt
pandas==2.2.0 
numpy==1.26.4 
matplotlib==3.8.2 
seaborn==0.13.2 
sqlalchemy==2.0.25 
jupyter==1.0.0 
scikit-learn==1.4.0 
scipy==1.12.0 
plotly==5.18.0 
pyyaml==6.0.1 
tqdm==4.66.1 
tabulate==0.9.0 
